Finish self_multiplier but still have free-memory issue
I have finished the function, and make them work properly. There are two key reasons why
I spent so many time on this function

1. I didn't make the length of the result in the strToInt long enough to do calculation in the later
section which resulted two differnt outputs on mac and linux. The reason why it's so important is
that once you didn't make the length enough, you can't manipulate the number out of the original
range. And it also made the output of adder_to_it_self wired, which casued by the reason above.

2. In the one digit_multiplier, I forgot to add result[i] to temp, which might not be 0 since you
you might have calculated a number which is more than 10 when i = i - 1.

Working on it, but still need to take care of the memory free part after finish this question.